+++ hutool-core/src/main/java/cn/hutool/core/collection/CollUtil.java
@@ -387,6 +387,19 @@ public class CollUtil {
 	public static <T> HashSet<T> newHashSet(T... ts) {
 		return newHashSet(false, ts);
 	}
+	
+	/**
+	 * 新建一个LinkedHashSet
+	 * 
+	 * @param <T> 集合元素类型
+	 * @param ts 元素数组
+	 * @return HashSet对象
+	 * @since 4.1.10
+	 */
+	@SafeVarargs
+	public static <T> LinkedHashSet<T> newLinkedHashSet(T... ts) {
+		return (LinkedHashSet<T>)newHashSet(true, ts);
+	}
 
 	/**
 	 * 新建一个HashSet
@@ -1741,7 +1754,6 @@ public class CollUtil {
 	 * @return 元素值
 	 * @since 4.0.6
 	 */
-	@SuppressWarnings("unchecked")
 	public static <T> T get(Collection<T> collection, int index) {
 		if (index < 0) {
 			index += collection.size();
@@ -1750,8 +1762,21 @@ public class CollUtil {
 			final List<T> list = ((List<T>) collection);
 			return list.get(index);
 		} else {
-			return (T) ((Collection<T>) collection).toArray()[index];
+			int i = 0;
+			for(T t : collection) {
+				if(i > index) {
+					break;
+				}else if(i == index) {
+					return t;
+				}
+				i++;
+			}
+			//检查越界
+			if(index >= i) {
+				throw new IndexOutOfBoundsException(StrUtil.format("Length is {} but index is {}", i, index));
+			}
 		}
+		return null;
 	}
 
 	/**
@@ -1811,6 +1836,18 @@ public class CollUtil {
 	 */
 	public static <T> T getFirst(Iterator<T> iterator) {
 		return IterUtil.getFirst(iterator);
+	}
+	
+	/**
+	 * 获取集合的最后一个元素
+	 * 
+	 * @param <T> 集合元素类型
+	 * @param collection {@link Collection}
+	 * @return 最后一个元素
+	 * @since 4.1.10
+	 */
+	public static <T> T getLast(Collection<T> collection) {
+		return get(collection, -1);
 	}
 
 	/**