+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### 新特性
 * 【extra】         Template增加Jfinal的Enjoy模板支持
 * 【core】          Assert增加checkBetween方法，Validator增加isBetween和validatorBetween
+* 【core】          增加CollUtil.getLast方法（感谢@【帝都】宁静）
 
 ### Bug修复
 * 【core】         修复BOMInputStream构造的问题（pr#22@Gitee）